+++ src/styles/Card.module.css
@@ -13,6 +13,14 @@
   border-radius: 8px;
 }
 
+@media screen and (max-width: 768px) {
+  .card {
+    flex-direction: column;
+    align-items: center;
+    width: 250px;
+  }
+}
+
 .img {
   max-width: 200px;
 }
