+++ src/styles/Header.module.css
@@ -6,11 +6,25 @@
   column-gap: 50px;
 }
 
+@media screen and (max-width: 768px) {
+  .container {
+    flex-direction: column;
+    row-gap: 15px;
+  }
+}
+
 .logo {
   max-width: 300px;
 }
 
+@media screen and (max-width: 768px) {
+  .logo {
+    max-width: 250px;
+  }
+}
+
 .input__container {
+  display: flex;
   border: 2px solid #02afc5;
   border-radius: 5px;
 }
