+++ src/types/appTypes.ts
@@ -1,0 +1,5 @@
+export interface AppProps {}
+
+export interface AppState {
+  searchValue: string;
+}
