+++ src/types/cardTypes.ts
@@ -1,0 +1,9 @@
+import { Data } from './dataTypes';
+
+export interface CardProps {
+  userInput: string | null;
+}
+
+export interface CardState {
+  characters: Data[];
+}
