+++ src/types/serchTypes.ts
@@ -1,0 +1,7 @@
+export interface SearchBarProps {
+  handleOnSearch: (userValue: string) => void;
+}
+
+export interface SearchBarState {
+  userInput: string | null;
+}
